Finished separation of current contents
Separation of classes into individual files is now complete. Current files are labeled accordingly for contents.
